--- guides/micronaut-hibernate-reactive/java/src/main/java/com/bloidonia/Application.java
@@ -1,9 +1,0 @@
-package com.bloidonia;
-
-import io.micronaut.runtime.Micronaut;
-
-public class Application {
-    public static void main(String[] args) {
-        Micronaut.run(Application.class, args);
-    }
-}
