+++ guides/micronaut-hibernate-reactive/java/src/main/java/example/micronaut/GenreController.java
@@ -1,4 +1,4 @@
-package com.bloidonia;
+package example.micronaut;
 
 import io.micronaut.http.annotation.Controller;
 import io.micronaut.http.annotation.Get;
