+++ guides/micronaut-hibernate-reactive/java/src/main/java/example/micronaut/GenreRepository.java
@@ -1,4 +1,4 @@
-package com.bloidonia;
+package example.micronaut;
 
 import org.reactivestreams.Publisher;
 import reactor.core.publisher.Mono;
