+++ guides/micronaut-hibernate-reactive/java/src/main/java/example/micronaut/GenreRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.bloidonia;
+package example.micronaut;
 
 import jakarta.inject.Singleton;
 import org.hibernate.SessionFactory;
